Move MavenResolver related classes to their own package
