+++ core/src/main/java/com/nikodoko/javaimports/resolver/maven/MavenDependencyFinder.java
@@ -1,4 +1,4 @@
-package com.nikodoko.javaimports.resolver;
+package com.nikodoko.javaimports.resolver.maven;
 
 import java.io.IOException;
 import java.io.Reader;
