+++ core/src/test/java/com/nikodoko/javaimports/resolver/maven/MavenDependencyResolverTest.java
@@ -1,9 +1,10 @@
-package com.nikodoko.javaimports.resolver;
+package com.nikodoko.javaimports.resolver.maven;
 
 import static com.google.common.truth.Truth.assertThat;
 
 import com.google.common.collect.ImmutableList;
 import com.nikodoko.javaimports.parser.Import;
+import com.nikodoko.javaimports.resolver.ImportWithDistance;
 import java.net.URL;
 import java.nio.file.Path;
 import java.nio.file.Paths;
